+++ pkg/botman/botman.go
@@ -26,6 +26,7 @@ type (
 		BotEndpointCoverageReport
 		BotManagementSetting
 		ChallengeAction
+		ChallengeInjectionRules
 		ChallengeInterceptionRules
 		ClientSideSecurity
 		ConditionalAction