+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # EDGEGRID GOLANG RELEASE NOTES
 
+## 7.1.0 (July 25, 2023)
+
+### FEATURES/ENHANCEMENTS:
+
+* APPSEC
+  * Added Bot Management API Support
+    * Challenge Injection Rules - read, update
+    * Add `CreateSecurityPolicyWithDefaultProtections` method to the `SecurityPolicy` interface to support creating a security policy with all available protections enabled.
+  * Update marshaling of PII learning setting
+
+### Deprecations
+
+* Challenge Interceptions Rules has been deprecated
+* Deprecate the following interfaces used to maintain individual policy protections:
+  * `ApiConstraintsProtection`
+  * `IPGeoProtection`
+  * `MalwareProtection`
+  * `NetworkLayerProtection`
+  * `RateProtection`
+  * `ReputationProtection`
+  * `SlowPostProtection`
+  * `WAFProtection`
+* Deprecate the `CreateSecurityPolicy` method of the `SecurityPolicy` interface.
+
 ## 7.0.0 (June 20, 2023)
 
 ### BREAKING CHANGES: